Rename views dir to view
