+++ packages/web/src/modules/register/RegisterConnector.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { RegisterView } from './views/RegisterView';
+import { RegisterView } from './view/RegisterView';
 import { RegisterController } from '@abb/controller';
 
 export class RegisterConnector extends React.PureComponent {
